Added basic FIPS test to the new simple top level
