+++ src/enc/aes_128_top_enc.vhdl
@@ -69,7 +69,7 @@ begin
                 end if;
 
                 if input_valid = '1' then
-                    if xor_init_vec = '1' then
+                    if xor_init_vec = '1' or init_vec_valid = '1' then
                         -- this will override the previous cipherblock
                         s_box_bus_in <= input_bus xor init_vec xor e_key(0); -- xor with initial vector
                         xor_init_vec_done <= '1'; -- Pulsed
